requirements.txt
# ===== Core =====
python-dotenv==1.0.1
requests>=2.32.3
numpy>=1.26.4,<2.0.0
scikit-learn>=1.4.2
tqdm>=4.66.4
regex>=2024.5.15
chardet>=5.2.0
python-dotenv==1.0.1

# ===== Document parsing =====
pypdf>=4.2.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
html2text>=2024.2.26

# ===== OpenAI =====
openai>=1.51.0
httpx>=0.27.0

# ===== Cross Encoder rerank/HF envirment=====
sentence-transformers==3.0.1
transformers>=4.44.0
huggingface_hub>=0.24.0
safetensors>=0.4.3
torch==2.4.1       

# ===== BM25 =====
rank-bm25>=0.2.2
